requirements.txt
ray[tune]
apache-airflow-providers-http
filelock
numpy
xgboost
xgboost_ray
modin
sklearn
pyarrow
#airflow-provider-ray
git+https://github.com/fletchjeff/airflow-provider-ray
